create promptRemove components specific for user roles deletion
